[wip] remote webgraph trait structure
