+++ Cargo.toml
@@ -30,7 +30,7 @@ debug = true
 [workspace.dependencies]
 ahash = "0.8.11"
 aho-corasick = "1.0"
-anyhow = {version = "1.0.72", features = ["backtrace"]}
+anyhow = { version = "1.0.72", features = ["backtrace"] }
 arc-swap = "1.5.0"
 async-channel = "1.8.0"
 async-stream = "0.3.3"
@@ -38,66 +38,66 @@ async-trait = "0.1"
 axum = "0.7.2"
 axum-macros = "0.4.0"
 base64 = "0.22.0"
-bincode = {git = "https://github.com/bincode-org/bincode", rev = "100685b", features = [
+bincode = { git = "https://github.com/bincode-org/bincode", rev = "100685b", features = [
   "serde",
-]}
+] }
 bindgen = "0.69.2"
 binggan = "0.8.0"
 bitflags = "2.5.0"
-bitpacking = {version = "0.9.2", default-features = false, features = [
+bitpacking = { version = "0.9.2", default-features = false, features = [
   "bitpacker4x",
   "bitpacker1x",
-]}
-bitvec = {version = "1.0.1", features = ["serde"]}
+] }
+bitvec = { version = "1.0.1", features = ["serde"] }
 bytecount = "0.6.7"
-bytemuck = {version = "1.13.1", features = ["derive"]}
+bytemuck = { version = "1.13.1", features = ["derive"] }
 byteorder = "1.4.3"
 candle-core = "0.3.3"
 candle-nn = "0.3.3"
 candle-transformers = "0.3.3"
-cc = {version = "1", features = ["parallel"]}
+cc = { version = "1", features = ["parallel"] }
 census = "0.4.2"
 chardetng = "0.1.17"
 chitchat = "0.5.0"
-chrono = {version = "0.4.23", features = ["serde"]}
-clap = {version = "4.4.6", features = ["derive"]}
+chrono = { version = "0.4.23", features = ["serde"] }
+clap = { version = "4.4.6", features = ["derive"] }
 cmake = "0.1"
-common = {path = "crates/common"}
+common = { path = "crates/common" }
 crc32fast = "1.3.2"
 criterion = "0.5.1"
 crossbeam-channel = "0.5.6"
 csv = "1.1.6"
-dashmap = {version = "5.4.0", features = ["rayon"]}
+dashmap = { version = "5.4.0", features = ["rayon"] }
 deadpool = "0.12.1"
 dirs = "5.0.1"
 downcast-rs = "1.2.0"
 encoding_rs = "0.8.31"
 enum_dispatch = "0.3.12"
 eventsource-stream = "0.2.3"
-fail = {version = "0.5.0", features = ["failpoints"]}
+fail = { version = "0.5.0", features = ["failpoints"] }
 fastdivide = "0.4.0"
 fend-core = "1.4.8"
 flate2 = "1.0.28"
 fnv = "1.0.3"
-fs4 = {version = "0.8.0"}
-fst = {version = "0.4.7", features = ["levenshtein"]}
+fs4 = { version = "0.8.0" }
+fst = { version = "0.4.7", features = ["levenshtein"] }
 futures = "0.3.21"
-futures-util = {version = "0.3.28"}
-half = {version = "2.2.1", features = ["serde"]}
-hashbrown = {version = "0.14.0", features = ["serde"]}
+futures-util = { version = "0.3.28" }
+half = { version = "2.2.1", features = ["serde"] }
+hashbrown = { version = "0.14.0", features = ["serde"] }
 htmlescape = "0.3.1"
 http = "1.0.0"
 image = "0.25.1"
-indicatif = {version = "0.17.7", features = ["rayon"]}
+indicatif = { version = "0.17.7", features = ["rayon"] }
 insta = "1.31"
 itertools = "0.13.0"
 js-sys = "0.3.60"
-lalrpop = {version = "0.20.0", features = ["lexer"]}
-lalrpop-util = {version = "0.20.0", features = ["lexer"]}
-lending-iter = {path = "crates/lending-iter"}
+lalrpop = { version = "0.20.0", features = ["lexer"] }
+lalrpop-util = { version = "0.20.0", features = ["lexer"] }
+lending-iter = { path = "crates/lending-iter" }
 levenshtein_automata = "0.2.1"
 libc = "0.2.142"
-log = {version = "0.4", features = ["release_max_level_info"]}
+log = { version = "0.4", features = ["release_max_level_info"] }
 logos = "0.13.0"
 lru = "0.12.0"
 lz-str = "0.2.1"
@@ -116,15 +116,19 @@ nom = "7.1.3"
 num-bigint = "0.4.1"
 num-rational = "0.4.1"
 oneshot = "0.1.7"
-openraft = {version = "0.9.1", features = [
+openraft = { version = "0.9.1", features = [
   "storage-v2",
   "serde",
   "tracing-log",
-]}
-ownedbytes = {path = "crates/ownedbytes"}
+] }
+ownedbytes = { path = "crates/ownedbytes" }
 paste = "1.0.11"
 percent-encoding = "2.3.0"
-postcard = {version = "1.0.8", features = ["experimental-derive", "alloc", "use-std"]}
+postcard = { version = "1.0.8", features = [
+  "experimental-derive",
+  "alloc",
+  "use-std",
+] }
 pretty_assertions = "1.2.1"
 proptest = "1.2.0"
 proptest-derive = "0.4.0"
@@ -134,54 +138,54 @@ rand = "0.8.5"
 rand_distr = "0.4.3"
 rayon = "1.5.3"
 redb = "2.0.0"
-regex = {version = "1.6.0", features = ["std", "unicode"]}
-reqwest = {version = "0.11.16", features = ["blocking", "stream", "json"]}
+regex = { version = "1.6.0", features = ["std", "unicode"] }
+reqwest = { version = "0.11.16", features = ["blocking", "stream", "json"] }
 ring = "0.17.3"
 rio_api = "0.8.4"
 rio_turtle = "0.8.4"
-rkyv = {version = "0.7.44", features = ["validation", "bytecheck"]}
-rusqlite = {version = "0.29.0", features = [
+rkyv = { version = "0.7.44", features = ["validation", "bytecheck"] }
+rusqlite = { version = "0.29.0", features = [
   "bundled",
   "modern-full",
   "buildtime_bindgen",
-]}
-rust-s3 = {version = "0.34.0", features = ["blocking", "tokio"]}
+] }
+rust-s3 = { version = "0.34.0", features = ["blocking", "tokio"] }
 rust-stemmers = "1.2.0"
 rustc-hash = "1.1"
 safetensors = "0.3.1"
-scylla = {version = "0.12.0", features = ["chrono"]}
-serde = {version = "1.0.137", features = ["rc", "derive"]}
+scylla = { version = "0.12.0", features = ["chrono"] }
+serde = { version = "1.0.137", features = ["rc", "derive"] }
 serde-wasm-bindgen = "0.6.3"
 serde_json = "1.0.81"
 serde_urlencoded = "0.7.1"
 smallvec = "1.8.0"
 stable_deref_trait = "1.2.0"
-strum = {version = "0.26.2", features = ["derive"]}
-tantivy = {path = "crates/tantivy"}
-tempfile = {version = "3.3.0"}
+strum = { version = "0.26.2", features = ["derive"] }
+tantivy = { path = "crates/tantivy" }
+tempfile = { version = "3.3.0" }
 thiserror = "1.0.31"
 tikv-jemallocator = "0.5"
-time = {version = "0.3.10", features = ["serde-well-known", "macros"]}
+time = { version = "0.3.10", features = ["serde-well-known", "macros"] }
 tokenizers = "0.13.2"
-tokio = {version = "1.23.1", features = ["full"]}
+tokio = { version = "1.23.1", features = ["full"] }
 tokio-stream = "0.1.11"
 toml = "0.8.2"
-tower = {version = "0.4.13", features = ["limit"]}
-tower-http = {version = "0.5.0", features = ["compression-gzip", "cors"]}
-tracing = {version = "0.1.34", features = ["release_max_level_info"]}
-tracing-subscriber = {version = "0.3.11", features = ["env-filter"]}
+tower = { version = "0.4.13", features = ["limit"] }
+tower-http = { version = "0.5.0", features = ["compression-gzip", "cors"] }
+tracing = { version = "0.1.34", features = ["release_max_level_info"] }
+tracing-subscriber = { version = "0.3.11", features = ["env-filter"] }
 tracing-test = "0.2.4"
 unicode-normalization = "0.1.23"
 unicode-segmentation = "1.11.0"
-url = {version = "2.4.0", features = ["serde"]}
-utoipa = {version = "4.2.3", features = ["axum_extras"]}
-utoipa-swagger-ui = {version = "7.0.0", features = ["axum"]}
-uuid = {version = "1.1.2", features = ["serde", "v4"]}
+url = { version = "2.4.0", features = ["serde"] }
+utoipa = { version = "4.2.3", features = ["axum_extras"] }
+utoipa-swagger-ui = { version = "7.0.0", features = ["axum"] }
+uuid = { version = "1.1.2", features = ["serde", "v4"] }
 wasm-bindgen = "0.2.83"
-whatlang = {version = "0.16.0", features = ["serde"]}
-xxhash-rust = {version = "0.8.10", features = ["xxh3", "const_xxh3"]}
+whatlang = { version = "0.16.0", features = ["serde"] }
+xxhash-rust = { version = "0.8.10", features = ["xxh3", "const_xxh3"] }
 zipf = "7.0.0"
-zstd = {version = "0.13", features = ["experimental"]}
+zstd = { version = "0.13", features = ["experimental"] }
 
 [profile.test.package]
 flate2.opt-level = 3
