+++ crates/core/src/entrypoint/ampc/approximated_harmonic_centrality/mapper.rs
@@ -20,8 +20,11 @@ use std::{cmp, collections::BTreeMap};
 
 use super::{worker::ApproxCentralityWorker, ApproxCentralityJob, Mapper};
 use crate::{
-    ampc::dht::upsert, entrypoint::ampc::approximated_harmonic_centrality::DhtTable,
-    kahan_sum::KahanSum, webgraph, webpage::html::links::RelFlags,
+    ampc::dht::upsert,
+    entrypoint::ampc::approximated_harmonic_centrality::DhtTable,
+    kahan_sum::KahanSum,
+    webgraph::{self, EdgeLimit},
+    webpage::html::links::RelFlags,
 };
 use rayon::prelude::*;
 
@@ -70,7 +73,11 @@ impl Workers {
             for outgoing in self
                 .worker
                 .graph()
-                .raw_outgoing_edges(&node, webgraph::EdgeLimit::Limit(MAX_OUTGOING_EDGES))
+                .search(
+                    &webgraph::query::ForwardlinksQuery::new(node)
+                        .with_limit(EdgeLimit::Limit(MAX_OUTGOING_EDGES)),
+                )
+                .unwrap_or_default()
                 .into_iter()
                 .filter(|e| !e.rel_flags.intersects(*SKIPPED_REL))
                 .map(|e| e.to)
