+++ crates/core/src/webgraph/shortest_path.rs
@@ -19,7 +19,7 @@ use std::{
     collections::{BTreeMap, BinaryHeap},
 };
 
-use super::{Node, NodeID, SmallEdge, Webgraph};
+use super::{query, EdgeLimit, Node, NodeID, SmallEdge, Webgraph};
 
 pub trait ShortestPaths {
     fn distances(&self, source: Node) -> BTreeMap<Node, u8>;
@@ -83,14 +83,21 @@ impl ShortestPaths for Webgraph {
     fn distances(&self, source: Node) -> BTreeMap<Node, u8> {
         self.raw_distances(source.id())
             .into_iter()
-            .filter_map(|(id, dist)| self.host_id2node(&id).unwrap().map(|node| (node, dist)))
+            .filter_map(|(id, dist)| {
+                self.search(&query::Id2NodeQuery::Host(id))
+                    .unwrap()
+                    .map(|node| (node, dist))
+            })
             .collect()
     }
 
     fn raw_distances_with_max(&self, source: NodeID, max_dist: u8) -> BTreeMap<NodeID, u8> {
         dijkstra_multi(
             &[source],
-            |node| self.raw_outgoing_edges(&node, super::EdgeLimit::Unlimited),
+            |node| {
+                self.search(&query::ForwardlinksQuery::new(node).with_limit(EdgeLimit::Unlimited))
+                    .unwrap_or_default()
+            },
             |edge| edge.to,
             Some(max_dist),
         )
@@ -99,7 +106,10 @@ impl ShortestPaths for Webgraph {
     fn raw_distances(&self, source: NodeID) -> BTreeMap<NodeID, u8> {
         dijkstra_multi(
             &[source],
-            |node| self.raw_outgoing_edges(&node, super::EdgeLimit::Unlimited),
+            |node| {
+                self.search(&query::ForwardlinksQuery::new(node).with_limit(EdgeLimit::Unlimited))
+                    .unwrap_or_default()
+            },
             |edge| edge.to,
             None,
         )
@@ -108,7 +118,10 @@ impl ShortestPaths for Webgraph {
     fn raw_reversed_distances(&self, source: NodeID) -> BTreeMap<NodeID, u8> {
         dijkstra_multi(
             &[source],
-            |node| self.raw_ingoing_edges(&node, super::EdgeLimit::Unlimited),
+            |node| {
+                self.search(&query::BacklinksQuery::new(node).with_limit(EdgeLimit::Unlimited))
+                    .unwrap_or_default()
+            },
             |edge| edge.from,
             None,
         )
@@ -121,7 +134,10 @@ impl ShortestPaths for Webgraph {
     ) -> BTreeMap<NodeID, u8> {
         dijkstra_multi(
             &[source],
-            |node| self.raw_ingoing_edges(&node, super::EdgeLimit::Unlimited),
+            |node| {
+                self.search(&query::BacklinksQuery::new(node).with_limit(EdgeLimit::Unlimited))
+                    .unwrap_or_default()
+            },
             |edge| edge.from,
             Some(max_dist),
         )
@@ -130,7 +146,11 @@ impl ShortestPaths for Webgraph {
     fn reversed_distances(&self, source: Node) -> BTreeMap<Node, u8> {
         self.raw_reversed_distances(source.id())
             .into_iter()
-            .filter_map(|(id, dist)| self.host_id2node(&id).unwrap().map(|node| (node, dist)))
+            .filter_map(|(id, dist)| {
+                self.search(&query::Id2NodeQuery::Host(id))
+                    .unwrap()
+                    .map(|node| (node, dist))
+            })
             .collect()
     }
 }
