+++ crates/tantivy/Cargo.toml
@@ -19,25 +19,26 @@ async-trait.workspace = true
 base64.workspace = true
 bitpacking.workspace = true
 byteorder.workspace = true
+bincode.workspace = true
 census.workspace = true
 crc32fast.workspace = true
 crossbeam-channel.workspace = true
 downcast-rs.workspace = true
-fail = {workspace = true, optional = true}
+fail = { workspace = true, optional = true }
 fastdivide.workspace = true
 fnv.workspace = true
-fs4 = {workspace = true, optional = true}
+fs4 = { workspace = true, optional = true }
 fst.workspace = true
-futures-util = {workspace = true, optional = true}
+futures-util = { workspace = true, optional = true }
 htmlescape.workspace = true
 itertools.workspace = true
 lending-iter.workspace = true
 levenshtein_automata.workspace = true
 log.workspace = true
 lru.workspace = true
-lz4_flex = {workspace = true, optional = true}
+lz4_flex = { workspace = true, optional = true }
 measure_time.workspace = true
-memmap2 = {workspace = true, optional = true}
+memmap2 = { workspace = true, optional = true }
 murmurhash32.workspace = true
 nom.workspace = true
 oneshot.workspace = true
@@ -50,7 +51,7 @@ serde.workspace = true
 serde_json.workspace = true
 smallvec.workspace = true
 strum.workspace = true
-tempfile = {workspace = true, optional = true}
+tempfile = { workspace = true, optional = true }
 thiserror.workspace = true
 time.workspace = true
 uuid.workspace = true
